requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
pymongo
email-validator
